4-ChartBlast_from_the_past/requirements.txt
requests~=2.28.1
spotipy~=2.23.0
Flask==2.3.2
beautifulsoup4~=4.12.2
python-dotenv==0.21.1
pip~=23.1.2
attrs~=22.1.0
cryptography==41.0.1
Pillow~=9.5.0
Jinja2~=3.1.2
MarkupSafe~=2.1.1
Werkzeug==2.2.3
click~=8.1.3
blinker~=1.6.2
itsdangerous~=2.1.2
setuptools~=65.5.0
visitor~=0.1.3
packaging~=23.1
zipp~=3.4.1
platformdirs~=3.2.0